# Refactoring #1: 