--- a/Pipfile
+++ b/Pipfile
@@ -34,6 +34,7 @@ twine = "*"
 pymonad = "*"
 distributed = "*"
 sparkpickle = "*"
+dask = "*"
 
 [packages]
 pytorch-lifestream = {editable = true, path = "."}