--- a/Pipfile
+++ b/Pipfile
@@ -40,4 +40,4 @@ dask = "*"
 pytorch-lifestream = {editable = true, path = "."}
 
 [requires]
-python_version = ">=3.8"
+python_version = "3.8"