--- a/Pipfile
+++ b/Pipfile
@@ -20,7 +20,7 @@ tqdm = ">=4.32.2"
 tensorboard = ">=2.1.0"
 pyhocon = "*"
 hydra-optuna-sweeper = '>=1.2.0'
-duckdb = "*"
+duckdb = "==0.8.0"
 pyarrow = '>=7.0'
 transformers = "*"
 hydra-core = '>=1.1.2'
@@ -32,13 +32,9 @@ scipy = "~=1.10.1"
 pymonad = "*"
 distributed = "*"
 dask = "*"
-torch = "==1.12.1"
-duckdb = "==0.8.0"
 
 [packages]
 pytorch-lifestream = {editable = true, path = "."}
-
-
 
 [requires]
 python_version = "3.8"