--- a/demo/finetune_examples/hparam_tuning/bin/data-preprocessing.sh
+++ b/demo/finetune_examples/hparam_tuning/bin/data-preprocessing.sh
@@ -1,9 +1,0 @@
-#!/usr/bin/env bash
-
-spark-submit \
-    --master local[8] \
-    --name "Gender Make Dataset" \
-    --driver-memory 16G \
-    --conf spark.driver.memoryOverhead="4g" \
-    --conf spark.local.dir="data/.spark_local_dir" \
-    data_preprocessing.py