--- a/demo/update-nb.sh
+++ b/demo/update-nb.sh
@@ -1,4 +1,0 @@
-jupyter nbconvert --execute --inplace coles-emb.ipynb
-jupyter nbconvert --execute --inplace pyspark-parquet.ipynb
-jupyter nbconvert --execute --inplace coles-catboost.ipynb
-jupyter nbconvert --execute --inplace coles-finetune.ipynb