--- a/ptls/data_load/data_module/coles_data_module.py
+++ b/ptls/data_load/data_module/coles_data_module.py
@@ -18,21 +18,19 @@ import warnings
 
 import pytorch_lightning as pl
 import torch
-
 from torch.utils.data import DataLoader
 from tqdm.auto import tqdm
 
 from ptls.data_load import IterableChain, padded_collate_wo_target
 from ptls.data_load.augmentations.build_augmentations import build_augmentations
-from ptls.data_load.iterable_processing.category_size_clip import CategorySizeClip
-from ptls.data_load.iterable_processing.feature_filter import FeatureFilter
-from ptls.data_load.iterable_processing.iterable_shuffle import IterableShuffle
-from ptls.data_load.iterable_processing.seq_len_filter import SeqLenFilter
-from ptls.data_load.list_splitter import ListSplitter
 from ptls.data_load.datasets.parquet_dataset import ParquetDataset, ParquetFiles
+from ptls.data_load.iterable_processing.feature_filter import FeatureFilter
+from ptls.data_load.iterable_processing.filter_collection import SeqLenFilter, CategorySizeClip
+from ptls.data_load.iterable_processing.iterable_shuffle import IterableShuffle
+from ptls.data_load.list_splitter import ListSplitter
 from ptls.data_load.partitioned_dataset import PartitionedDataset, PartitionedDataFiles
-from ptls.metric_learn.dataset import nested_list_to_flat_with_collate
 from ptls.frames.coles import split_strategy
+from ptls.metric_learn.dataset import nested_list_to_flat_with_collate
 from ptls.metric_learn.dataset.splitting_dataset import IterableSplittingDataset, MapSplittingDataset
 
 logger = logging.getLogger(__name__)
@@ -136,6 +134,7 @@ class ColesDataModuleTrain(pl.LightningDataModule):
             raise AttributeError(f'Unknown split strategy: {self.setup_conf.split_by}')
 
     def build_iterable_processing(self, part):
+
         if part == 'train':
             yield SeqLenFilter(min_seq_len=self.train_conf.min_seq_len)
         yield FeatureFilter(keep_feature_names=self.category_names)