--- a/ptls/data_load/data_module/cpc_data_module.py
+++ b/ptls/data_load/data_module/cpc_data_module.py
@@ -12,21 +12,18 @@ client_2          -> client_2_features, client_2_id
 import logging
 import warnings
 
-import numpy as np
 import pytorch_lightning as pl
-
 from torch.utils.data import DataLoader
 from tqdm.auto import tqdm
 
 from ptls.data_load import IterableAugmentations, IterableChain, padded_collate_wo_target
 from ptls.data_load.augmentations.build_augmentations import build_augmentations
 from ptls.data_load.data_module.map_augmentation_dataset import MapAugmentationDataset
-from ptls.data_load.iterable_processing.category_size_clip import CategorySizeClip
-from ptls.data_load.iterable_processing.feature_filter import FeatureFilter
-from ptls.data_load.iterable_processing.iterable_shuffle import IterableShuffle
-from ptls.data_load.iterable_processing.seq_len_filter import SeqLenFilter
-from ptls.data_load.list_splitter import ListSplitter
 from ptls.data_load.datasets.parquet_dataset import ParquetDataset, ParquetFiles
+from ptls.data_load.iterable_processing.feature_filter import FeatureFilter
+from ptls.data_load.iterable_processing.filter_collection import CategorySizeClip, SeqLenFilter
+from ptls.data_load.iterable_processing.iterable_shuffle import IterableShuffle
+from ptls.data_load.list_splitter import ListSplitter
 from ptls.data_load.partitioned_dataset import PartitionedDataset, PartitionedDataFiles
 
 logger = logging.getLogger(__name__)