--- a/ptls/data_load/data_module/rtd_data_module.py
+++ b/ptls/data_load/data_module/rtd_data_module.py
@@ -21,14 +21,13 @@ from tqdm.auto import tqdm
 from ptls.data_load import IterableAugmentations, IterableChain, padded_collate_wo_target
 from ptls.data_load.augmentations.build_augmentations import build_augmentations
 from ptls.data_load.data_module.map_augmentation_dataset import MapAugmentationDataset
-from ptls.data_load.iterable_processing.category_size_clip import CategorySizeClip
-from ptls.data_load.iterable_processing.feature_filter import FeatureFilter
-from ptls.data_load.iterable_processing.iterable_shuffle import IterableShuffle
-from ptls.data_load.iterable_processing.seq_len_filter import SeqLenFilter
-from ptls.data_load.list_splitter import ListSplitter
 from ptls.data_load.datasets.parquet_dataset import ParquetDataset, ParquetFiles
-from ptls.data_load.partitioned_dataset import PartitionedDataset, PartitionedDataFiles
+from ptls.data_load.iterable_processing.feature_filter import FeatureFilter
+from ptls.data_load.iterable_processing.filter_collection import CategorySizeClip, SeqLenFilter
+from ptls.data_load.iterable_processing.iterable_shuffle import IterableShuffle
+from ptls.data_load.list_splitter import ListSplitter
 from ptls.data_load.padded_batch import PaddedBatch
+from ptls.data_load.partitioned_dataset import PartitionedDataset, PartitionedDataFiles
 
 logger = logging.getLogger(__name__)