--- a/ptls/data_load/datasets/duckdb_dataset.py
+++ b/ptls/data_load/datasets/duckdb_dataset.py
@@ -1,4 +1,5 @@
 from typing import Iterable, List
+
 import duckdb
 import torch
 
@@ -51,7 +52,7 @@ class DuckDbDataset(torch.utils.data.IterableDataset):
     
         rel = duckdb.sql(query)
 
-        while(True):
+        while True:
             recs = rel.fetchmany(1000)
             if not recs:
                 break