… .github/workflows/main.yml
… .github/workflows/unit_test.yml
… Pipfile.lock
… README.md
… demo/CoLES-multimodal.ipynb
… demo/classification_examples/coles-catboost.ipynb
… demo/coles-emb.ipynb
… demo/coles-finetune.ipynb
… demo/coles-pretrained-embeddings.ipynb
… demo/embedding_examples/coles-emb.ipynb
… demo/embedding_examples/event-sequence-local-embeddings.ipynb
… demo/embedding_examples/mlm-emb.ipynb
… demo/embedding_examples/nsp-sop-emb.ipynb
… demo/embedding_examples/tabformer-emb.ipynb
… demo/extended-inference.ipynb
… demo/finetune_examples/coles-embedder-finetune-custom-lr.ipynb
… demo/finetune_examples/coles-embedder-finetune.ipynb
… demo/finetune_examples/coles-finetune.ipynb
… demo/finetune_examples/hparam_tuning/README.md
… demo/finetune_examples/hparam_tuning/conf/data_preprocessing.yaml
… demo/finetune_examples/hparam_tuning/conf/data_preprocessing/default.yaml
… demo/finetune_examples/hparam_tuning/conf/one_layer_head.yaml
… demo/finetune_examples/hparam_tuning/conf/simple_config.yaml
… demo/finetune_examples/hparam_tuning/tuning.py
… demo/hydra-CoLES-training.ipynb
… demo/multilabel-classification.ipynb
… demo/preprocessing-demo.ipynb
… demo/preprocessing_examples/preprocessing-demo.ipynb
… demo/preprocessing_examples/pyspark-parquet.ipynb
… demo/preprocessing_examples/supervised-sequence-to-target-transformer.ipynb