**Key changes:**

- Parallelization of preprocessing using the dask library
The preprocessing time on coles_emb demo was reduced from ~40 seconds to ~25 seconds. CPU: Ryzen 5600x.
[Profiling](https://i.imgur.com/RRMIqN9.png)

- Refactoring the preprocessing module
The preprocessing time on the coles_emb demo was reduced from ~25 seconds to ~10 seconds. CPU: Ryzen 5600x.
[Profiling](https://i.imgur.com/EXe0VVT.png)

- Refactoring of data loading modules
Improved code readability, optimized default parameters experimentally on alphabattle and gender scenarios (GPU: A6000, 3080), fixed bugs.

- Encoder refactoring
Improved code readability, compatibility with new changes, optimization of some functions.

- A number of filters became a single class (Seqlen, CategorySizeClip, DeleteNan, IdFilter, ISeqLenLimit, FilterNonArray, IdFilterDf)

- A number of preprocessing classes also moved to processing collection (FeatureBinScaler, FeatureFilter, FeatureRename, FeatureTypeCast) and became a part of a single one

- Some dependencies were updated

- Unit-tests were updated

- Added docstrings and typehints

- Minor PEP8 fixes in general. **Major syntax fixes will be applied after pull-request approve**.